Minor bugs solved
