+++ src/app.ts
@@ -12,7 +12,7 @@ import { conductorRouter } from './conductor/conductor.routes.js'
 import { cargaRouter } from './carga/carga.routes.js'
 import { licenciaRouter } from './licencia/licencia.routes.js'
 import { estadoTrenRouter } from './estadoTren/estadoTren.routes.js'
-import { catRouter } from './categoriaDenuncia/categoriaDenunica.routes.js'
+import { catRouter } from './categoriaDenuncia/categoriaDenuncia.routes.js'
 import { observacionRouter } from './observacion/observacion.routes.js'
 import { lineaCargaRouter } from './lineaCarga/lineaCarga.routes.js'
 import { viajeRouter } from './viaje/viaje.routes.js'
